split settings
